+++ src/main/java/server/re_ser.java
@@ -11,7 +11,7 @@ public interface re_ser {
 	 input getInputById(int id);
 	 void updateinput(input obj);
 	 void deleteinput(int id);
-	 List<input> findbynumber(int numb);
+	 List<input> findbynumber(long numb);
 	 List<input>getpending();
 	 List<input> accept();
